requirements.txt
datasets
spacy
wordfreq
lemminflect
pyyaml
wn
ucimlrepo
pandas
